[Refactor] separate `RecyclableClient` class from `HiveMetaClient`
